fixed issues
