+++ bimex-frontend/src/components/ListaProyectos.jsx
@@ -2,6 +2,7 @@ import { useState, useEffect, useRef } from "react";
 import { useTranslation } from "react-i18next";
 import { obtenerTodosLosProyectos, stroopsAMXNe } from "../stellar/contrato";
 import { parsearError } from "../utils/errores.js";
+import CardProyecto, { SkeletonCard } from "./CardProyecto";
 
 export default function ListaProyectos({ onSeleccionar, onCrear, refrescar }) {
   const { t } = useTranslation();
@@ -220,25 +221,6 @@ export default function ListaProyectos({ onSeleccionar, onCrear, refrescar }) {
   );
 }
 
-// ── Skeleton card ──────────────────────────────────────────────────────────────
-function SkeletonCard() {
-  return (
-    <article className="card" aria-hidden="true" style={{ ...estilos.card, pointerEvents: 'none', userSelect: 'none' }}>
-      <div style={estilos.cardTop}>
-        <div className="skeleton" style={{ height: 22, width: 90, borderRadius: 99 }} />
-      </div>
-      <div className="skeleton" style={{ height: 20, width: '70%', marginBottom: 8 }} />
-      <div className="skeleton" style={{ height: 14, width: '40%', marginBottom: 18 }} />
-      <div className="skeleton" style={{ height: 8, borderRadius: 4, marginBottom: 8 }} />
-      <div style={estilos.statsRow}>
-        <div className="skeleton" style={{ height: 14, width: 70 }} />
-        <div className="skeleton" style={{ height: 14, width: 70 }} />
-      </div>
-      <div className="skeleton" style={{ height: 36, marginTop: 16, borderRadius: 6 }} />
-    </article>
-  );
-}
-
 // ── Stat strip item ──────────────────────────────────────────────────────────
 function StatStrip({ label, valor, mono, highlight }) {
   return (
@@ -253,105 +235,6 @@ function StatStrip({ label, valor, mono, highlight }) {
       }}>
         {valor}
       </div>
-    </div>
-  );
-}
-
-// ── Config de estado ─────────────────────────────────────────────────────────
-const ESTADO_CFG = {
-  EtapaInicial: { badge: "badge-muted",  btnLabelKey: "card.contributeBtn", btnClass: "btn-secondary" },
-  EnProgreso:   { badge: "badge-teal",   btnLabelKey: "card.contributeBtn", btnClass: "btn-secondary" },
-  Liberado:     { badge: "badge-amber",  btnLabelKey: "card.detailBtn",     btnClass: "btn-secondary" },
-  Abandonado:   { badge: "badge-red",    btnLabelKey: "card.takeControlBtn",btnClass: "btn-ghost"     },
-};
-
-// ── Card ─────────────────────────────────────────────────────────────────────
-function CardProyecto({ proyecto, onClick }) {
-  const { t } = useTranslation();
-  const meta     = Number(proyecto.meta);
-  const aportado = Number(proyecto.aportado);
-  const pct      = meta > 0 ? Math.min((aportado / meta) * 100, 100) : 0;
-  const estado   = proyecto.estado ?? "EtapaInicial";
-  const cfg      = ESTADO_CFG[estado] ?? ESTADO_CFG.EtapaInicial;
-  const btnLabel = t(cfg.btnLabelKey);
-
-  return (
-    <article
-      className="card"
-      role="listitem"
-      style={{ ...estilos.card, opacity: estado === "Abandonado" ? 0.75 : 1 }}
-      onClick={onClick}
-      aria-label={`${proyecto.nombre}, ${t(`status.${estado}`)}, ${pct.toFixed(0)}%`}
-    >
-      {/* Estado + verificado */}
-      <div style={estilos.cardTop}>
-        <span className={`badge ${cfg.badge}`}>
-          <span className="badge-dot" />
-          {t(`status.${estado}`)}
-        </span>
-        {proyecto.doc_hash && (
-          <span style={{
-            background: "var(--green-dim)", border: "1px solid rgba(22,163,74,0.20)",
-            color: "var(--green)", fontSize: "0.7rem", fontWeight: 600,
-            padding: "2px 8px", borderRadius: "99px",
-          }}>
-            Verificado
-          </span>
-        )}
-      </div>
-
-      {/* Nombre + dueño */}
-      <h3 style={estilos.nombre}>{proyecto.nombre}</h3>
-      <p style={{ fontSize: "0.75rem", color: "var(--subtle)", fontFamily: "'SFMono-Regular','Consolas',monospace", marginBottom: 0 }}>
-        {proyecto.dueno.slice(0, 6)}...{proyecto.dueno.slice(-4)}
-      </p>
-
-      {/* Progreso */}
-      <div style={{ marginTop: 18 }}>
-        <div style={{ display: "flex", justifyContent: "space-between", marginBottom: 6 }}>
-          <span style={{ fontSize: "0.78rem", color: "var(--muted)" }}>
-            {t("lista.funding")}
-          </span>
-          <span style={{ fontSize: "0.78rem", color: "var(--green)", fontWeight: 700 }}>
-            {pct.toFixed(0)}%
-          </span>
-        </div>
-        <div
-          className="progress-track"
-          role="progressbar"
-          aria-valuenow={Math.round(pct)}
-          aria-valuemin={0}
-          aria-valuemax={100}
-          aria-valuetext={`${pct.toFixed(0)}%`}
-        >
-          <div className="progress-fill" style={{ width: `${pct}%` }} />
-        </div>
-      </div>
-
-      {/* Stats */}
-      <div style={estilos.statsRow}>
-        <StatItem label={t("lista.locked")} valor={stroopsAMXNe(proyecto.aportado)} />
-        <StatItem label={t("lista.goal")}   valor={stroopsAMXNe(proyecto.meta)}     muted />
-      </div>
-
-      {/* CTA */}
-      <button
-        className={`btn ${cfg.btnClass}`}
-        style={{ width: "100%", marginTop: 16, justifyContent: "center" }}
-        onClick={(e) => { e.stopPropagation(); onClick(); }}
-        aria-label={`${btnLabel} ${proyecto.nombre}`}
-      >
-        {btnLabel}
-      </button>
-    </article>
-  );
-}
-
-function StatItem({ label, valor, muted }) {
-  return (
-    <div style={{ textAlign: "center" }}>
-      <div style={{ fontSize: "0.7rem", color: "var(--subtle)", textTransform: "uppercase", letterSpacing: "0.06em", fontWeight: 600 }}>{label}</div>
-      <div style={{ fontFamily: "'SFMono-Regular','Consolas',monospace", fontSize: "0.82rem", color: muted ? "var(--muted)" : "var(--text2)", marginTop: 3, fontWeight: 600 }}>{valor}</div>
     </div>
   );
 }
