+++ bimex-frontend/src/components/MiCuenta.jsx
@@ -1,151 +1,36 @@
-import { useState, useEffect, useRef } from "react";
+import { useState, useEffect } from "react";
 import { useTranslation } from "react-i18next";
-import { createClient } from "@supabase/supabase-js";
 import { parsearError } from "../utils/errores.js";
-import { usePaginacion, ControlPagina } from "../hooks/usePaginacion.js";
-import {
-  obtenerTodosLosProyectos,
-  obtenerAportacion,
-  calcularYield,
-  stroopsAMXNe,
-} from "../stellar/contrato";
+import { obtenerTodosLosProyectos, obtenerAportacion, calcularYield, stroopsAMXNe } from "../stellar/contrato";
+import TabMisProyectos      from "./TabMisProyectos";
+import TabMisContribuciones from "./TabMisContribuciones";
+import NotificacionesPanel  from "./NotificacionesPanel";
+import { SkeletonTableRows } from "./TabMisContribuciones";
 
-const supabase = import.meta.env.VITE_SUPABASE_URL
-  ? createClient(import.meta.env.VITE_SUPABASE_URL, import.meta.env.VITE_SUPABASE_ANON_KEY)
-  : null;
+// ─── Metro card icons ────────────────────────────────────────────────────────
 
-// ─── Config de estado ─────────────────────────────────────────────────────────
+const IconCapital = () => (
+  <svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="var(--navy)" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
+    <circle cx="12" cy="12" r="10"/>
+    <path d="M12 6v2m0 8v2M9.5 9.5a2.5 2.5 0 0 1 5 0c0 1.5-1.5 2-2.5 2.5S9.5 13 9.5 14.5a2.5 2.5 0 0 0 5 0"/>
+  </svg>
+);
 
-const ESTADO_CFG = {
-  EtapaInicial: { labelKey: "status.EtapaInicial", badgeClass: "badge-muted"  },
-  EnProgreso:   { labelKey: "status.EnProgreso",   badgeClass: "badge-teal"   },
-  Liberado:     { labelKey: "status.Liberado",     badgeClass: "badge-amber"  },
-  Abandonado:   { labelKey: "status.Abandonado",   badgeClass: "badge-red"    },
-  EnRevision:   { labelKey: "status.EnRevision",   badgeClass: null, customStyle: { background: "var(--amber-dim)", color: "var(--amber)", border: "1px solid rgba(217,119,6,0.20)" } },
-  Rechazado:    { labelKey: "status.Rechazado",    badgeClass: "badge-red"    },
-};
+const IconRendimiento = () => (
+  <svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="var(--green)" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
+    <polyline points="22 7 13.5 15.5 8.5 10.5 2 17"/>
+    <polyline points="16 7 22 7 22 13"/>
+  </svg>
+);
 
-function getBadgeCfg(estado) {
-  return ESTADO_CFG[estado] ?? ESTADO_CFG.EtapaInicial;
-}
+const IconProyectos = () => (
+  <svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="var(--navy)" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
+    <rect x="2" y="7" width="20" height="14" rx="2"/>
+    <path d="M16 7V5a2 2 0 0 0-2-2h-4a2 2 0 0 0-2 2v2"/>
+  </svg>
+);
 
-// ─── Helpers ──────────────────────────────────────────────────────────────────
-
-function pct(aportado, meta) {
-  const a = Number(aportado ?? 0);
-  const m = Number(meta ?? 0);
-  return m > 0 ? Math.min((a / m) * 100, 100) : 0;
-}
-
-function puedeRetirar(estado) {
-  return estado === "Liberado" || estado === "Abandonado";
-}
-
-// ─── SVG Icons ────────────────────────────────────────────────────────────────
-
-function IconCapital() {
-  return (
-    <svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="var(--navy)" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
-      <circle cx="12" cy="12" r="10"/>
-      <path d="M12 6v2m0 8v2M9.5 9.5a2.5 2.5 0 0 1 5 0c0 1.5-1.5 2-2.5 2.5S9.5 13 9.5 14.5a2.5 2.5 0 0 0 5 0"/>
-    </svg>
-  );
-}
-
-function IconRendimiento() {
-  return (
-    <svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="var(--green)" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
-      <polyline points="22 7 13.5 15.5 8.5 10.5 2 17"/>
-      <polyline points="16 7 22 7 22 13"/>
-    </svg>
-  );
-}
-
-function IconProyectos() {
-  return (
-    <svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="var(--navy)" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
-      <rect x="2" y="7" width="20" height="14" rx="2"/>
-      <path d="M16 7V5a2 2 0 0 0-2-2h-4a2 2 0 0 0-2 2v2"/>
-    </svg>
-  );
-}
-
-function IconFile() {
-  return (
-    <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
-      <path d="M14 2H6a2 2 0 0 0-2 2v16a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V8z"/>
-      <polyline points="14 2 14 8 20 8"/>
-    </svg>
-  );
-}
-
-function IconSeedling() {
-  return (
-    <svg width="40" height="40" viewBox="0 0 24 24" fill="none" stroke="var(--border2)" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
-      <path d="M7 20c0-5.5 5.5-9 9-9"/>
-      <path d="M2 9c0-3.3 2.7-6 6-6 3.3 0 6 2.7 6 6 0 4.4-3 6-6 6-2 0-4-.5-6-2"/>
-    </svg>
-  );
-}
-
-function IconInbox() {
-  return (
-    <svg width="40" height="40" viewBox="0 0 24 24" fill="none" stroke="var(--border2)" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
-      <polyline points="22 12 16 12 14 15 10 15 8 12 2 12"/>
-      <path d="M5.45 5.11 2 12v6a2 2 0 0 0 2 2h16a2 2 0 0 0 2-2v-6l-3.45-6.89A2 2 0 0 0 16.76 4H7.24a2 2 0 0 0-1.79 1.11z"/>
-    </svg>
-  );
-}
-
-function IconBell() {
-  return (
-    <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="var(--navy)" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
-      <path d="M18 8A6 6 0 0 0 6 8c0 7-3 9-3 9h18s-3-2-3-9"/>
-      <path d="M13.73 21a2 2 0 0 1-3.46 0"/>
-    </svg>
-  );
-}
-
-// ─── Sub-componentes ──────────────────────────────────────────────────────────
-
-function Spinner() {
-  const { t } = useTranslation();
-  return (
-    <div style={estilos.loadingWrap} role="status" aria-live="polite" aria-label={t("cuenta.loading")}>
-      <div style={estilos.spinner} aria-hidden="true" />
-      <p style={{ color: "var(--muted)", marginTop: 16, fontSize: "0.9rem" }}>{t("cuenta.loading")}</p>
-    </div>
-  );
-}
-
-function SkeletonTableRows({ count = 5 }) {
-  return (
-    <div aria-hidden="true" style={{ pointerEvents: 'none', userSelect: 'none' }}>
-      {Array.from({ length: count }).map((_, i) => (
-        <div key={i} style={{ display: 'flex', gap: 16, padding: '12px 14px', borderBottom: '1px solid var(--border)', alignItems: 'center' }}>
-          <div className="skeleton" style={{ height: 16, width: '25%' }} />
-          <div className="skeleton" style={{ height: 16, width: '10%' }} />
-          <div className="skeleton" style={{ height: 16, width: '12%' }} />
-          <div className="skeleton" style={{ height: 16, width: '12%' }} />
-          <div className="skeleton" style={{ height: 16, width: '15%' }} />
-          <div className="skeleton" style={{ height: 16, width: '10%' }} />
-          <div className="skeleton" style={{ height: 28, width: '16%', borderRadius: 6 }} />
-        </div>
-      ))}
-    </div>
-  );
-}
-
-function EstadoBadge({ estado }) {
-  const { t } = useTranslation();
-  const cfg = getBadgeCfg(estado);
-  if (cfg.customStyle) {
-    return <span className="badge" style={cfg.customStyle}>{t(cfg.labelKey)}</span>;
-  }
-  return <span className={`badge ${cfg.badgeClass}`}>{t(cfg.labelKey)}</span>;
-}
-
-// ─── Metric Cards (top summary) ───────────────────────────────────────────────
+// ─── Metric Cards ────────────────────────────────────────────────────────────────────
 
 function MetricCard({ icon, label, value, accent }) {
   return (
@@ -153,376 +38,6 @@ function MetricCard({ icon, label, value, accent }) {
       <div style={estilos.metricIcon}>{icon}</div>
       <div style={estilos.metricLabel}>{label}</div>
       <div style={{ ...estilos.metricValue, color: accent ?? "var(--navy)" }}>{value}</div>
-    </div>
-  );
-}
-
-// ─── Pestaña: Mis proyectos ───────────────────────────────────────────────────
-
-function CardMiProyecto({ proyecto, onVerProyecto }) {
-  const { t } = useTranslation();
-  const progreso = pct(proyecto.aportado, proyecto.meta);
-
-  return (
-    <article className="card" style={estilos.card}>
-      <div style={estilos.cardTop}>
-        <h3 style={estilos.cardTitulo}>{proyecto.nombre}</h3>
-        <EstadoBadge estado={proyecto.estado} />
-      </div>
-
-      <div style={{ marginTop: 14 }}>
-        <div style={{ display: "flex", justifyContent: "space-between", marginBottom: 6 }}>
-          <span style={{ fontSize: "0.76rem", color: "var(--muted)" }}>{t("cuenta.funding")}</span>
-          <span style={{ fontSize: "0.76rem", color: "var(--navy)", fontWeight: 700 }}>
-            {progreso.toFixed(0)}%
-          </span>
-        </div>
-        <div
-          className="progress-track"
-          role="progressbar"
-          aria-valuenow={Math.round(progreso)}
-          aria-valuemin={0}
-          aria-valuemax={100}
-          aria-valuetext={`${progreso.toFixed(0)}% del objetivo`}
-        >
-          <div className="progress-fill" style={{ width: `${progreso}%` }} />
-        </div>
-      </div>
-
-      <div style={estilos.statsRow}>
-        <div>
-          <div style={estilos.statLabel}>{t("cuenta.raised")}</div>
-          <div style={estilos.statValor}>{stroopsAMXNe(proyecto.aportado)}</div>
-        </div>
-        <div style={{ textAlign: "right" }}>
-          <div style={estilos.statLabel}>{t("cuenta.goal")}</div>
-          <div style={estilos.statValor}>{stroopsAMXNe(proyecto.meta)}</div>
-        </div>
-      </div>
-
-      <button
-        className="btn btn-secondary"
-        style={{ width: "100%", marginTop: 16, justifyContent: "center" }}
-        onClick={() => onVerProyecto(proyecto)}
-        aria-label={`${t("cuenta.viewDetails")} ${proyecto.nombre}`}
-      >
-        {t("cuenta.viewDetails")}
-      </button>
-    </article>
-  );
-}
-
-function TabMisProyectos({ proyectos, direccion, onVerProyecto, t }) {
-  const misProyectos = proyectos.filter((p) => p.dueno === direccion);
-  const gridRef = useRef(null);
-
-  const { datosPagina, pagina, setPagina, totalPaginas, total, cargandoPagina } =
-    usePaginacion(misProyectos, [proyectos, direccion]);
-
-  const handlePaginaChange = (nuevaPagina) => {
-    setPagina(nuevaPagina);
-    if (gridRef.current) {
-      gridRef.current.scrollIntoView({ behavior: "smooth", block: "start" });
-    }
-  };
-
-  if (misProyectos.length === 0) {
-    return (
-      <div style={estilos.empty}>
-        <IconSeedling />
-        <p style={{ fontSize: "1.05rem", fontWeight: 700, color: "var(--text)", marginTop: 16 }}>
-          {t("cuenta.noProjects")}
-        </p>
-        <p style={{ fontSize: "0.86rem", color: "var(--muted)", marginTop: 6 }}>
-          {t("cuenta.noProjectsHint")}
-        </p>
-      </div>
-    );
-  }
-
-  return (
-    <div ref={gridRef} style={{ opacity: cargandoPagina ? 0.5 : 1, transition: "opacity 0.15s" }}>
-      {cargandoPagina && (
-        <div style={{ display: "grid", gridTemplateColumns: "repeat(auto-fill, minmax(285px, 1fr))", gap: 20, marginBottom: 16 }}>
-          {Array.from({ length: Math.min(4, datosPagina.length || 4) }).map((_, i) => (
-            <div key={`skeleton-${i}`} className="card" style={{ padding: 20 }}>
-              <div className="skeleton" style={{ height: 16, width: "60%", borderRadius: 4, marginBottom: 12 }} />
-              <div className="skeleton" style={{ height: 8, width: "100%", borderRadius: 4 }} />
-              <div className="skeleton" style={{ height: 16, width: "40%", borderRadius: 4, marginTop: 14 }} />
-            </div>
-          ))}
-        </div>
-      )}
-      {!cargandoPagina && (
-        <div className="cuenta-grid" style={estilos.grid} role="list" aria-label={t("cuenta.ariaProjects")}>
-          {datosPagina.map((p) => (
-            <CardMiProyecto key={p.id} proyecto={p} onVerProyecto={onVerProyecto} />
-          ))}
-        </div>
-      )}
-      <ControlPagina pagina={pagina} totalPaginas={totalPaginas} onChange={handlePaginaChange} t={t} />
-    </div>
-  );
-}
-
-// ─── Pestaña: Mis contribuciones ──────────────────────────────────────────────
-
-function TabMisContribuciones({ proyectos, direccion, onVerProyecto }) {
-  const { t } = useTranslation();
-  const [contribuciones, setContribuciones] = useState([]);
-  const [cargando, setCargando] = useState(true);
-  const [errorContrib, setErrorContrib] = useState(null);
-  const tableRef = useRef(null);
-
-  useEffect(() => {
-    if (proyectos.length === 0) {
-      setCargando(false);
-      return;
-    }
-
-    async function cargarContribuciones() {
-      setCargando(true);
-      setErrorContrib(null);
-      try {
-        const resultados = await Promise.all(
-          proyectos.map(async (p) => {
-            const [aportacion, yieldAcum] = await Promise.all([
-              obtenerAportacion(p.id, direccion),
-              calcularYield(p.id, direccion),
-            ]);
-            return { proyecto: p, aportacion, yieldAcum };
-          })
-        );
-        setContribuciones(resultados.filter((r) => r.aportacion > BigInt(0)));
-      } catch (e) {
-        setErrorContrib(parsearError(e));
-      } finally {
-        setCargando(false);
-      }
-    }
-
-    cargarContribuciones();
-  }, [proyectos, direccion]);
-
-  // Hook must be called unconditionally (Rules of Hooks)
-  const {
-    datosPagina,
-    pagina,
-    setPagina,
-    totalPaginas,
-    cargandoPagina,
-  } = usePaginacion(contribuciones, [contribuciones]);
-
-  const handlePaginaChange = (nuevaPagina) => {
-    setPagina(nuevaPagina);
-    if (tableRef.current) {
-      tableRef.current.scrollIntoView({ behavior: "smooth", block: "start" });
-    }
-  };
-
-  if (cargando) {
-    return (
-      <div role="status" aria-live="polite" aria-label={t("cuenta.loading")}>
-        <SkeletonTableRows count={5} />
-      </div>
-    );
-  }
-
-  if (errorContrib) {
-    return (
-      <div role="alert" style={{ color: "var(--error, #DC2626)", background: "rgba(220,38,38,0.06)", border: "1px solid rgba(220,38,38,0.18)", borderRadius: "var(--radius-sm)", padding: "14px 18px", fontSize: "0.86rem", marginTop: 8 }}>
-        {errorContrib}
-      </div>
-    );
-  }
-
-  if (contribuciones.length === 0) {
-    return (
-      <div style={estilos.empty}>
-        <IconInbox />
-        <p style={{ fontSize: "1.05rem", fontWeight: 700, color: "var(--text)", marginTop: 16 }}>
-          {t("cuenta.noContributions")}
-        </p>
-        <p style={{ fontSize: "0.86rem", color: "var(--muted)", marginTop: 6 }}>
-          {t("cuenta.noContributionsHint")}
-        </p>
-      </div>
-    );
-  }
-
-  return (
-    <div ref={tableRef} style={{ overflowX: "auto", opacity: cargandoPagina ? 0.5 : 1, transition: "opacity 0.15s" }}>
-      {cargandoPagina && <SkeletonTableRows count={5} />}
-      {!cargandoPagina && (
-        <table style={estilos.table}>
-          <thead>
-            <tr>
-              <th style={estilos.th}>{t("cuenta.colProyecto")}</th>
-              <th style={estilos.th}>{t("cuenta.colModo")}</th>
-              <th style={{ ...estilos.th, textAlign: "right" }}>{t("cuenta.colCapital")}</th>
-              <th style={{ ...estilos.th, textAlign: "right" }}>{t("cuenta.colRendimiento")}</th>
-              <th style={estilos.th}>{t("cuenta.colEstado")}</th>
-              <th style={estilos.th}>{t("cuenta.colCierre")}</th>
-              <th style={estilos.th} />
-            </tr>
-          </thead>
-          <tbody>
-            {datosPagina.map(({ proyecto, aportacion, yieldAcum }) => {
-              const puedeRet = puedeRetirar(proyecto.estado);
-              const modo = proyecto.modo ?? "Inversor";
-              return (
-                <tr key={proyecto.id} style={estilos.tr}>
-                  <td style={estilos.td}>
-                    <span style={{ fontWeight: 600, color: "var(--text)" }}>{proyecto.nombre}</span>
-                  </td>
-                  <td style={estilos.td}>
-                    <span className={modo === "Mecenas" ? "badge badge-teal" : "badge badge-navy"}>
-                      {modo}
-                    </span>
-                  </td>
-                  <td style={{ ...estilos.td, textAlign: "right", fontVariantNumeric: "tabular-nums" }}>
-                    {stroopsAMXNe(aportacion)}
-                  </td>
-                  <td style={{ ...estilos.td, textAlign: "right", color: "var(--green)", fontWeight: 600, fontVariantNumeric: "tabular-nums" }}>
-                    {stroopsAMXNe(yieldAcum)}
-                  </td>
-                  <td style={estilos.td}>
-                    <EstadoBadge estado={proyecto.estado} />
-                  </td>
-                  <td style={{ ...estilos.td, color: "var(--muted)", fontSize: "0.83rem" }}>
-                    {proyecto.fecha_cierre ?? "—"}
-                  </td>
-                  <td style={{ ...estilos.td, textAlign: "right" }}>
-                    <div style={{ display: "flex", gap: 6, justifyContent: "flex-end" }}>
-                      <button
-                        className="btn btn-secondary"
-                        style={{ padding: "6px 12px", fontSize: "0.8rem" }}
-                        onClick={() => onVerProyecto(proyecto)}
-                        aria-label={`${t("cuenta.viewDetailsShort")} ${proyecto.nombre}`}
-                      >
-                        <IconFile />
-                      </button>
-                      {puedeRet && (
-                        <button
-                          className="btn btn-amber"
-                          style={{ padding: "6px 12px", fontSize: "0.8rem" }}
-                          onClick={() => onVerProyecto(proyecto)}
-                          aria-label={`${t("cuenta.withdraw")} ${proyecto.nombre}`}
-                        >
-                          {t("cuenta.withdraw")}
-                        </button>
-                      )}
-                    </div>
-                  </td>
-                </tr>
-              );
-            })}
-          </tbody>
-        </table>
-      )}
-      <ControlPagina pagina={pagina} totalPaginas={totalPaginas} onChange={handlePaginaChange} t={t} />
-    </div>
-  );
-}
-
-// ─── Notificaciones ───────────────────────────────────────────────────────────
-
-function NotificacionesPanel({ direccion }) {
-  const [email,   setEmail]   = useState("");
-  const [enabled, setEnabled] = useState(true);
-  const [estado,  setEstado]  = useState("idle"); // idle | saving | ok | error
-  const [cargado, setCargado] = useState(false);
-
-  useEffect(() => {
-    if (!direccion || !supabase) return;
-    supabase
-      .from("user_notifications")
-      .select("email, notifications_enabled")
-      .eq("wallet_address", direccion)
-      .maybeSingle()
-      .then(({ data }) => {
-        if (data) { setEmail(data.email); setEnabled(data.notifications_enabled); }
-        setCargado(true);
-      })
-      .catch(() => setCargado(true));
-  }, [direccion]);
-
-  async function guardar(e) {
-    e.preventDefault();
-    if (!email || !supabase) return;
-    setEstado("saving");
-    const { error } = await supabase
-      .from("user_notifications")
-      .upsert({ wallet_address: direccion, email, notifications_enabled: enabled }, { onConflict: "wallet_address" });
-    setEstado(error ? "error" : "ok");
-    setTimeout(() => setEstado("idle"), 3000);
-  }
-
-  if (!cargado) return null;
-
-  return (
-    <div style={estilos.notiPanel}>
-      <div style={{ display: "flex", alignItems: "flex-start", justifyContent: "space-between", gap: 16, marginBottom: 14 }}>
-        <div style={{ display: "flex", alignItems: "center", gap: 10 }}>
-          <IconBell />
-          <div>
-            <div style={{ fontWeight: 700, fontSize: "0.9rem", color: "var(--text)" }}>
-              Notificaciones por email
-            </div>
-            <div style={{ fontSize: "0.76rem", color: "var(--muted)", marginTop: 2 }}>
-              Recibe alertas cuando tu proyecto sea aprobado, financiado o tenga yield disponible.
-            </div>
-          </div>
-        </div>
-        <button
-          role="switch"
-          aria-checked={enabled}
-          onClick={() => setEnabled(v => !v)}
-          style={{
-            width: 44, height: 24, borderRadius: 99, border: "none", cursor: "pointer",
-            background: enabled ? "var(--navy)" : "var(--border2)",
-            position: "relative", flexShrink: 0, transition: "background 0.2s",
-          }}
-          aria-label={enabled ? "Desactivar notificaciones" : "Activar notificaciones"}
-        >
-          <span style={{
-            position: "absolute", top: 3,
-            left: enabled ? 22 : 3,
-            width: 18, height: 18, borderRadius: "50%",
-            background: "#fff", transition: "left 0.2s",
-          }} />
-        </button>
-      </div>
-
-      {enabled && (
-        <form onSubmit={guardar} style={{ display: "flex", gap: 8 }}>
-          <input
-            type="email"
-            required
-            placeholder="tu@email.com"
-            value={email}
-            onChange={e => setEmail(e.target.value)}
-            style={{
-              flex: 1, padding: "9px 14px",
-              borderRadius: "var(--radius-sm)",
-              border: "1.5px solid var(--border2)",
-              fontFamily: "inherit", fontSize: "0.88rem",
-              outline: "none", color: "var(--text)",
-            }}
-            onFocus={e => { e.target.style.borderColor = "var(--navy)"; }}
-            onBlur={e => { e.target.style.borderColor = "var(--border2)"; }}
-            aria-label="Email para notificaciones"
-          />
-          <button
-            type="submit"
-            disabled={estado === "saving"}
-            className="btn btn-primary"
-            style={{ whiteSpace: "nowrap", padding: "9px 18px" }}
-          >
-            {estado === "saving" ? "…" : estado === "ok" ? "Guardado" : estado === "error" ? "Error" : "Guardar"}
-          </button>
-        </form>
-      )}
     </div>
   );
 }
@@ -777,9 +292,7 @@ const estilos = {
     padding: "20px 22px",
     boxShadow: "var(--shadow-sm)",
   },
-  metricIcon: {
-    marginBottom: 10,
-  },
+  metricIcon: { marginBottom: 10 },
   metricLabel: {
     fontSize: "0.7rem",
     color: "var(--muted)",
@@ -792,16 +305,6 @@ const estilos = {
     fontSize: "1.35rem",
     fontWeight: 800,
     fontVariantNumeric: "tabular-nums",
-  },
-
-  // Notification panel
-  notiPanel: {
-    background: "var(--card)",
-    border: "1.5px solid var(--border)",
-    borderRadius: "var(--radius)",
-    padding: "20px 24px",
-    marginBottom: 28,
-    boxShadow: "var(--shadow-sm)",
   },
 
   // Tabs
@@ -843,103 +346,4 @@ const estilos = {
     fontWeight: 700,
   },
 
-  // Grid de cards (proyectos)
-  grid: {
-    display: "grid",
-    gridTemplateColumns: "repeat(auto-fill, minmax(285px, 1fr))",
-    gap: 20,
-  },
-
-  // Card individual
-  card: {
-    display: "flex",
-    flexDirection: "column",
-    cursor: "default",
-  },
-  cardTop: {
-    display: "flex",
-    justifyContent: "space-between",
-    alignItems: "flex-start",
-    gap: 10,
-    marginBottom: 4,
-  },
-  cardTitulo: {
-    fontSize: "1rem",
-    fontWeight: 700,
-    color: "var(--text)",
-    lineHeight: 1.3,
-    flex: 1,
-  },
-  statsRow: {
-    display: "flex",
-    justifyContent: "space-between",
-    marginTop: 16,
-    paddingTop: 14,
-    borderTop: "1.5px solid var(--border)",
-  },
-  statLabel: {
-    fontSize: "0.70rem",
-    color: "var(--muted)",
-    textTransform: "uppercase",
-    letterSpacing: "0.06em",
-    fontWeight: 700,
-    marginBottom: 3,
-  },
-  statValor: {
-    fontSize: "0.84rem",
-    color: "var(--text2)",
-    fontWeight: 600,
-    fontVariantNumeric: "tabular-nums",
-  },
-
-  // Contributions table
-  table: {
-    width: "100%",
-    borderCollapse: "collapse",
-    fontSize: "0.88rem",
-  },
-  th: {
-    padding: "10px 14px",
-    textAlign: "left",
-    fontSize: "0.70rem",
-    textTransform: "uppercase",
-    letterSpacing: "0.07em",
-    fontWeight: 700,
-    color: "var(--muted)",
-    borderBottom: "2px solid var(--border)",
-    whiteSpace: "nowrap",
-  },
-  tr: {
-    borderBottom: "1px solid var(--border)",
-  },
-  td: {
-    padding: "12px 14px",
-    color: "var(--text2)",
-    verticalAlign: "middle",
-  },
-
-  // Estado vacío
-  empty: {
-    display: "flex",
-    flexDirection: "column",
-    alignItems: "center",
-    padding: "80px 0",
-    textAlign: "center",
-  },
-
-  // Carga
-  loadingWrap: {
-    display: "flex",
-    flexDirection: "column",
-    alignItems: "center",
-    padding: "80px 0",
-  },
-  spinner: {
-    width: 36,
-    height: 36,
-    border: "3px solid var(--navy-dim)",
-    borderTopColor: "var(--navy)",
-    borderRadius: "50%",
-    animation: "spin 0.8s linear infinite",
-  },
 };
