+++ bimex-frontend/src/hooks/usePaginacion.jsx
